src/main/webapp/app/entities/form/update/form-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { FormService } from '../service/form.service';
import { IForm, Form } from '../form.model';

import { FormUpdateComponent } from './form-update.component';

describe('Form Management Update Component', () => {
  let comp: FormUpdateComponent;
  let fixture: ComponentFixture<FormUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let formService: FormService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [FormUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(FormUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(FormUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    formService = TestBed.inject(FormService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const form: IForm = { id: 456 };

      activatedRoute.data = of({ form });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(form));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Form>>();
      const form = { id: 123 };
      jest.spyOn(formService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ form });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: form }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(formService.update).toHaveBeenCalledWith(form);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Form>>();
      const form = new Form();
      jest.spyOn(formService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ form });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: form }));
      saveSubject.complete();

      // THEN
      expect(formService.create).toHaveBeenCalledWith(form);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Form>>();
      const form = { id: 123 };
      jest.spyOn(formService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ form });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(formService.update).toHaveBeenCalledWith(form);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
